requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
boto3==1.37.17
botocore==1.37.17
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.1.7
cms==0.0.2
colorama==0.4.6
cryptography==46.0.2
cssbeautifier==1.15.1
cssselect2==0.7.0
distlib==0.3.9
dj-database-url==2.2.0
Django==5.2.6
django-admin-sortable2==2.2.4
django-allauth==65.12.0
django-classy-tags==4.1.0
django-cms==5.0.3
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-entangled==0.6.2
django-environ==0.11.2
django-filer==3.3.1
django-formtools==2.5.1
django-fsm==2.8.2
django-fsm-2==4.1.0
django-js-asset==3.1.2
django-parler==2.3
django-polymorphic==3.1.0
django-recaptcha==4.1.0
django-sekizai==4.1.0
django-storages==1.14.6
django-treebeard==4.7.1
django-upgrade==1.25.0
djangocms-admin-style==3.3.1
djangocms-alias==3.0.2
djangocms-attributes-field==4.0.0
djangocms-file==3.0.1
djangocms-frontend==2.1.2
djangocms-googlemap==2.2.0
djangocms-link==5.1.0
djangocms-picture==4.1.1
djangocms-snippet==3.1.1
djangocms-style==3.1.0
djangocms-text==0.9.3
djangocms-versioning==2.5.0
djangocms-video==3.1.0
djlint==1.34.1
easy-thumbnails==2.10
EditorConfig==0.12.4
filelock==3.18.0
html-tag-names==0.1.2
html-void-elements==0.1.0
html5lib==1.1
identify==2.6.12
idna==3.11
jmespath==1.0.1
jsbeautifier==1.15.1
json5==0.9.25
lxml==5.2.2
nh3==0.3.0
nodeenv==1.9.1
oauthlib==3.3.1
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.8
pre_commit==4.2.0
psycopg==3.2.6
psycopg-binary==3.2.6
pycparser==2.23
PyJWT==2.10.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
reportlab==4.2.2
requests==2.32.5
s3transfer==0.11.4
setuptools==80.8.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
svglib==1.5.1
tinycss2==1.3.0
tokenize_rt==6.2.0
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.3.0
virtualenv==20.31.2
webencodings==0.5.1
wheel==0.46.3
whitenoise==6.7.0
